Move encrypt action classes to their own package
