+++ x-pack/plugin/enterprise-search/src/main/java/org/elasticsearch/xpack/enterprisesearch/action/encrypt/EncryptResponse.java
@@ -7,7 +7,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.enterprisesearch.action;
+package org.elasticsearch.xpack.enterprisesearch.action.encrypt;
 
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.update.UpdateResponse;
