+++ x-pack/plugin/enterprise-search/src/main/java/org/elasticsearch/xpack/enterprisesearch/action/encrypt/EncryptRestHandler.java
@@ -5,23 +5,16 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.enterprisesearch.action;
+package org.elasticsearch.xpack.enterprisesearch.action.encrypt;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.action.update.UpdateResponse;
 import org.elasticsearch.client.internal.node.NodeClient;
 import org.elasticsearch.rest.BaseRestHandler;
 import org.elasticsearch.rest.RestRequest;
-import org.elasticsearch.rest.RestResponse;
-import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.rest.action.RestStatusToXContentListener;
-import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentType;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
